deploy: revert back to git push + polling
In #19 we removed the polling script to rely on the clever cli's
ability to wait on `clever deploy` command.

After more than 6 months of experimentation it seems the command is
still not reliable (sometimes the command never returns, sometimes it
returns even when the deployment has finished..).

Thus we are reverting back to our manual process of polling `clever
activity` every 5 seconds.

The timeout for this polling is set at the Ansible level (not in the
bash script) and set for 40 minutes. This should give enough time for
any of our applications to deploy.