deploy: try to restart the app when the CLI asks to do it
In some cases (rollback or deployment errored on clever side),
clever-tools will fail when calling `clever deploy` because it thinks
the currently being deployed commit is already deployed. However in
some cases the git pushed commit is not the currently running commit.

Leading to a failure in our deployment process. The proposed solution
from the clever tools error is to launch `clever restart`.

Until we wait for
https://github.com/CleverCloud/clever-tools/issues/422 to be fixed,
let's do what the error message say: try to restart in case of a
deployment failure.